requirements.txt
tensorflow==2.17.1
keras==2.17.0
streamlit==1.32.2
numpy==1.24.4
pillow==10.2.0
matplotlib==3.7.5
scikit-learn==1.3.2
seaborn==0.13.2
joblib==1.3.2
